feat: begin Billing Info screen #298
