+++ frontend/lib/models/billing.dart
@@ -6,7 +6,11 @@ class BillingStatus {
   final bool receipt;
   final bool paid;
 
-  BillingStatus({required this.proForma, required this.invoice, required this.receipt, required this.paid});
+  BillingStatus(
+      {required this.proForma,
+      required this.invoice,
+      required this.receipt,
+      required this.paid});
 
   factory BillingStatus.fromJson(Map<String, dynamic> json) {
     return BillingStatus(
@@ -22,7 +26,7 @@ class BillingStatus {
         'invoice': invoice,
         'receipt': receipt,
         'paid': paid,
-  };
+      };
 
   @override
   String toString() {
@@ -41,12 +45,21 @@ class Billing {
   final String notes;
   final bool visible;
 
-  Billing({required this.id, required this.status, required this.event, this.company, required this.value, required this.invoiceNumber, required this.emission, required this.notes, required this.visible});
+  Billing(
+      {required this.id,
+      required this.status,
+      required this.event,
+      this.company,
+      required this.value,
+      required this.invoiceNumber,
+      required this.emission,
+      required this.notes,
+      required this.visible});
 
   factory Billing.fromJson(Map<String, dynamic> json) {
     return Billing(
       id: json['id'],
-      status:BillingStatus.fromJson(json['status']),
+      status: BillingStatus.fromJson(json['status']),
       event: json['event'],
       company: json['company'],
       value: json['value'],
@@ -56,7 +69,6 @@ class Billing {
       visible: json['visible'],
     );
   }
-
 
   Map<String, dynamic> toJson() => {
         'id': id,
